# docs/openapi/farm-service-v1.yaml
openapi: 3.1.0
info:
  title: ACM Farm Service API
  description: API documentation for ACM Farm Service
  license:
    name: Apache 2.0
    url: https://springdoc.org
  version: 1.0.0
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/farms/{farmId}/certification/scope:
    put:
      tags:
      - certification-controller
      operationId: updateScope
      parameters:
      - name: farmId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCertificationScopesRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListCertificationScopeResponse"
  /api/v1/farms/{farmId}/certification/items/{itemId}:
    put:
      tags:
      - certification-controller
      operationId: updateItemStatus
      parameters:
      - name: farmId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: itemId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCertificationItemRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
...
  /api/v1/farms/{farmId}/certification/export-dossier:
    post:
      tags:
      - certification-controller
      operationId: exportDossier
      parameters:
      - name: farmId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: X-User-Id
        in: header
        required: false
        schema:
          type: integer
          format: int64
      - name: diaryClientProvider
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/ObjectProviderSeasonProductionDiaryClient"
      - name: documentServiceProvider
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/ObjectProviderFarmDocumentService"
      - name: currentUserServiceProvider
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/ObjectProviderCurrentUserService"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportDossierRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseFarmDocumentResponse"
...
  /api/v1/farms/{farmId}/certification/apply:
    post:
      tags:
      - certification-controller
      operationId: applyCertification
      parameters:
      - name: farmId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
...
  /api/v1/farms/{farmId}/certification:
    get:
      tags:
      - certification-controller
      operationId: getCertificationDetails
      parameters:
      - name: farmId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCertificationDetailsResponse"
...
components:
  schemas:
...
    ScopeItem:
      type: object
      properties:
        seasonId:
          type: integer
          format: int32
        registeredAreaHa:
          type: number
          minimum: 0.0001
      required:
      - registeredAreaHa
      - seasonId
    UpdateCertificationScopesRequest:
      type: object
      properties:
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/ScopeItem"
          maxItems: 50
          minItems: 0
      required:
      - scopes
    ApiResponseListCertificationScopeResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          type: array
          items:
            $ref: "#/components/schemas/CertificationScopeResponse"
        data:
          type: array
          items:
            $ref: "#/components/schemas/CertificationScopeResponse"
        generatedAt:
          type: string
    CertificationScopeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        seasonId:
          type: integer
          format: int32
        plotId:
          type: integer
          format: int32
        plotName:
          type: string
        cropId:
          type: integer
          format: int32
        cropName:
          type: string
        varietyId:
          type: integer
          format: int32
        varietyName:
          type: string
        registeredAreaHa:
          type: number
        expectedYieldKg:
          type: number
    UpdateCertificationItemRequest:
      type: object
      properties:
        status:
          type: string
        evidenceUrl:
          type: string
        notes:
          type: string
    ApiResponseString:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          type: string
        data:
          type: string
        generatedAt:
          type: string
...
    ApiResponseFarmDocumentResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/FarmDocumentResponse"
        data:
          $ref: "#/components/schemas/FarmDocumentResponse"
        generatedAt:
          type: string
    FarmDocumentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        farmId:
          type: integer
          format: int32
        documentType:
          type: string
        documentTypeLabel:
          type: string
        title:
          type: string
        description:
          type: string
        fileUrl:
          type: string
        issuedDate:
          type: string
          format: date
        expiryDate:
          type: string
          format: date
        isExpired:
          type: boolean
        isExpiringSoon:
          type: boolean
        verificationStatus:
          type: string
        verifiedByName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ExportDossierRequest:
      type: object
      properties:
        seasonIds:
          type: array
          items:
            type: integer
            format: int32
    ObjectProviderSeasonProductionDiaryClient:
      type: object
      properties:
        object:
          $ref: "#/components/schemas/SeasonProductionDiaryClient"
        ifUnique:
          $ref: "#/components/schemas/SeasonProductionDiaryClient"
        ifAvailable:
          $ref: "#/components/schemas/SeasonProductionDiaryClient"
    SeasonProductionDiaryClient:
      type: object
    FarmDocumentService:
      type: object
    ObjectProviderFarmDocumentService:
      type: object
      properties:
        object:
          $ref: "#/components/schemas/FarmDocumentService"
        ifUnique:
          $ref: "#/components/schemas/FarmDocumentService"
        ifAvailable:
          $ref: "#/components/schemas/FarmDocumentService"
    CurrentUserService:
      type: object
      properties:
        currentUserId:
          type: integer
          format: int64
        currentRole:
          type: string
        currentUserFullName:
          type: string
    ObjectProviderCurrentUserService:
      type: object
      properties:
        object:
          $ref: "#/components/schemas/CurrentUserService"
        ifUnique:
          $ref: "#/components/schemas/CurrentUserService"
        ifAvailable:
          $ref: "#/components/schemas/CurrentUserService"
...
    MissingEvidenceItem:
      type: object
      properties:
        itemCode:
          type: string
        category:
          type: string
        description:
          type: string
...
    ApiResponseCertificationDetailsResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        code:
          type: string
        message:
          type: string
        result:
          $ref: "#/components/schemas/CertificationDetailsResponse"
        data:
          $ref: "#/components/schemas/CertificationDetailsResponse"
        generatedAt:
          type: string
    CertificationDetailsResponse:
      type: object
      properties:
        recordId:
          type: integer
          format: int32
        farmId:
          type: integer
          format: int32
        standardCode:
          type: string
        standardName:
          type: string
        complianceScore:
          type: number
        status:
          type: string
        appliedAt:
          type: string
          format: date-time
        certifiedAt:
          type: string
          format: date-time
        expiryDate:
          type: string
          format: date
        auditorNotes:
          type: string
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/CertificationScopeResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/CertificationItemDetail"
        isEligible:
          type: boolean
        certificateNumber:
          type: string
        nextPeriodicReviewDate:
          type: string
          format: date
        publishedAt:
          type: string
          format: date-time
        missingMandatoryEvidenceCount:
          type: integer
          format: int32
        missingEvidenceItems:
          type: array
          items:
            $ref: "#/components/schemas/MissingEvidenceItem"
    CertificationItemDetail:
      type: object
      properties:
        id:
          type: integer
          format: int32
        itemCode:
          type: string
        category:
          type: string
        description:
          type: string
        isMandatory:
          type: boolean
        weightPct:
          type: number
        dataSourceType:
          type: string
        dataSourceQuery:
          type: string
        status:
          type: string
        evidenceUrl:
          type: string
        notes:
          type: string
        checkedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT